requirement.txt
streamlit==1.51.0
mysql-connector-python==9.3.0
pandas==2.3.3
